# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swagger Uni - OpenAPI 3.0",
    "description": "This is a sample Uni Server based on the OpenAPI 3.0 specification.",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/admin/professor/all": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Professors",
        "description": "get a list of all professors",
        "operationId": "getProfessorsAdmin",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
...
    "/api/admin/professor/:id": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get One Professor",
        "description": "get one professor with given id",
        "operationId": "getOneProfessorAdmin",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "update professor",
        "description": "update professor with the information in req.body",
        "operationId": "updateProfessor",
        "requestBody": {
          "description": "enter new professor's info",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "user not found"
          },
          "500": {
            "description": "server error"
          }
        }
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "delete professor",
        "description": "update professor with given id",
        "operationId": "deleteProfessor",
        "responses": {
          "404": {
            "description": "user not found"
          },
          "500": {
            "description": "server error"
          }
        }
      }
    },
...
    "/api/professor/all": {
      "get": {
        "tags": [
          "manager"
        ],
        "summary": "Get Professors",
        "description": "get a list of all professors",
        "operationId": "getProfessors",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
...
    "/api/professor/:id": {
      "get": {
        "tags": [
          "manager"
        ],
        "summary": "Get One Professor",
        "description": "get one professor with given id",
        "operationId": "getOneProfessor",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "500": {
            "description": "Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "professor"
        ],
        "summary": "update professor",
        "description": "update professor with the information in req.body",
        "operationId": "updateProfessoAdmin",
        "requestBody": {
          "description": "enter new professor's info",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "user not found"
          },
          "500": {
            "description": "server error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "idNumber": {
            "type": "string",
            "example": "1234"
          },
          "password": {
            "type": "string",
            "example": "password111"
          },
          "email": {
            "type": "string",
            "example": "John@example.com"
          },
          "phonenumber": {
            "type": "string",
            "example": "555-555-5555"
          },
          "department": {
            "type": "string",
            "example": "English"
          },
          "role": {
            "type": "string",
            "enum": [
              "professor",
              "student",
              "manager",
              "admin"
            ]
          },
          "degree": {
            "type": "string",
            "example": "Bachelor's"
          },
          "year": {
            "type": "string",
            "example": "1"
          },
          "term": {
            "type": "string",
            "example": "Fall"
          },
          "average": {
            "type": "string",
            "example": "3.0"
          },
          "rank": {
            "type": "string",
            "example": "Professor"
          },
          "major": {
            "type": "string",
            "example": "Mathematics"
          }
        }
      },
...
    }
  }
}